Renamed apps 'users' to 'core' and 'polls' to 'pollish' - Renamed folders - Updated module imports - Updated migration dependencies and foregin key references
